+++ apps/test-app/src/main.tsx
@@ -12,7 +12,6 @@ import { routeTree } from "./routeTree.gen";
 const router = createRouter({
   routeTree,
   context: {
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     auth: undefined!,
   },
 });
@@ -25,7 +24,6 @@ declare module "@tanstack/react-router" {
 }
 
 // Render the app
-// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 const rootElement = document.getElementById("root")!;
 if (!rootElement.innerHTML) {
   const root = ReactDOM.createRoot(rootElement);
@@ -39,12 +37,11 @@ function InnerApp() {
 
 function ConditionalStrictMode({ children }: { children: React.ReactNode }) {
   const [strict, setStrict] = React.useState(() => {
-    return router.parseLocation().search.strict !== 0;
+    return router.state.location.search.strict !== 0;
   });
   React.useEffect(() => {
-    return router.subscribe("onLoad", (ctx) => {
-      const appParams = ctx.toLocation.search as { strict: 0 };
-      setStrict(appParams.strict !== 0);
+    return router.subscribe("onLoad", () => {
+      setStrict(router.state.location.search.strict !== 0);
     });
   }, []);
   if (!strict) {