+++ ui/appui-react/src/appui-react/layout/widget/widget-content-layout/header/HeaderIconToolbar.tsx
@@ -12,6 +12,8 @@ import {
   IconButton,
   MenuDivider,
   MenuItem,
+  Text,
+  Tooltip,
 } from "@itwin/itwinui-react";
 import { SvgMore, SvgSearch } from "@itwin/itwinui-icons-react";
 import { HeaderSearch } from "./HeaderSearch.js";
@@ -90,6 +92,21 @@ export function HeaderIconToolbar(props: HeaderIconToolbarProps) {
                     ) : (
                       <React.Fragment key={index}></React.Fragment>
                     )
+                  ) : icon.tooltipContent ? (
+                    <Tooltip content={icon.tooltipContent} key={index}>
+                      <MenuItem
+                        isSelected={icon.isActive}
+                        disabled={icon.disabled}
+                        key={index}
+                        onClick={() => {
+                          icon.onClick();
+                          close();
+                        }}
+                        startIcon={<>{icon.icon}</>}
+                      >
+                        {icon.label}
+                      </MenuItem>
+                    </Tooltip>
                   ) : (
                     <MenuItem
                       isSelected={icon.isActive}
@@ -134,7 +151,18 @@ export function HeaderIconToolbar(props: HeaderIconToolbarProps) {
           ) : (
             <IconButton
               disabled={icon.disabled}
-              label={icon.label}
+              label={
+                icon.tooltipContent ? (
+                  <div className="nz-widget-widgetContentLayout-header-headerIconToolbar-iconButton-tooltip">
+                    <Text variant="small">{icon.label}</Text>
+                    <Text isMuted variant="small">
+                      {icon.tooltipContent}
+                    </Text>
+                  </div>
+                ) : (
+                  icon.label
+                )
+              }
               key={icon.label}
               size={props.iconSize}
               styleType="borderless"