+++ ui/appui-react/src/appui-react/toolbar/ToolbarHelper.tsx
@@ -105,7 +105,6 @@ export class ToolbarHelper {
   public static getIconReactNode(
     item: ActionButton | GroupButton
   ): React.ReactNode {
-    // eslint-disable-next-line @typescript-eslint/no-deprecated
     return IconHelper.getIconReactNode(item.icon, item.internalData);
   }
 
@@ -132,8 +131,8 @@ export class ToolbarHelper {
       isHidden: itemDef.isHidden,
       isDisabled: itemDef.isDisabled,
       icon:
-        internalData?.get(IconHelper.reactIconKey) ?? // eslint-disable-line @typescript-eslint/no-deprecated
-        internalData?.get(IconHelper.conditionalIconItemKey) ?? // eslint-disable-line @typescript-eslint/no-deprecated
+        internalData?.get(IconHelper.reactIconKey) ??
+        internalData?.get(IconHelper.conditionalIconItemKey) ??
         icon ??
         itemDef.iconSpec,
       label: this.getStringOrConditionalString(itemDef.rawLabel),