+++ ui/appui-react/src/appui-react/toolsettings/InternalToolSettingsManager.ts
@@ -146,7 +146,7 @@ export class InternalToolSettingsManager {
   /** Get ToolSettings Properties sync event. */
   public static readonly onSyncToolSettingsProperties =
     new BeUiEvent<SyncToolSettingsPropertiesEventArgs>(); // eslint-disable-line @typescript-eslint/no-deprecated
-  // eslint-disable-next-line @typescript-eslint/no-deprecated
+
   public static readonly onReloadToolSettingsProperties = new BeUiEvent<void>();
 
   /** Gets the Id of the active tool. If a tool is not active, blank is returned.