+++ ui/imodel-components-react/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Change Log - @itwin/imodel-components-react
 
+## 5.24.2
+
+### Patch Changes
+
+- @itwin/components-react@5.24.2
+- @itwin/core-react@5.24.2
+
+## 5.24.1
+
+### Patch Changes
+
+- @itwin/components-react@5.24.1
+- @itwin/core-react@5.24.1
+
+## 5.24.0
+
+### Patch Changes
+
+- @itwin/components-react@5.24.0
+- @itwin/core-react@5.24.0
+
+## 5.23.1
+
+### Patch Changes
+
+- @itwin/components-react@5.23.1
+- @itwin/core-react@5.23.1
+
 ## 5.23.0
 
 ### Patch Changes